light: fix extended texinfo array type

+++ include/light/light.hh
@@ -379,7 +379,7 @@ extern int write_litfile;
 extern int write_luxfile;
 extern qboolean onlyents;
 extern qboolean scaledonly;
-extern uint32_t *extended_texinfo_flags;
+extern uint64_t *extended_texinfo_flags;
 extern qboolean novisapprox;
 extern bool nolights;
